// lib/api/generated/models/authSessionResponseData.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Heymoa API
 * Heymoa 서버 REST API
 * OpenAPI spec version: 1.0.0
 */
import type { AuthSessionResponseDataState } from "./authSessionResponseDataState";
import type { AuthSessionResponseDataUser } from "./authSessionResponseDataUser";

/**
 * 성공 응답 데이터
 */
export type AuthSessionResponseData = {
  /** 세션 상태 */
  state: AuthSessionResponseDataState;
  /** authenticated일 때만 현재 사용자 */
  user?: AuthSessionResponseDataUser;
};